# AUTOMATION/web/ui/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soulglade — API du studio",
    "description": "\nBackend local du studio Soulglade. Aucune authentification : le serveur\nn'accepte que ce qui vient de lui-même (voir `api/security.py`).\n\n**Conventions transverses**\n\n* presque toute route lit `?character=<id>` et le valide avant de toucher au\n  disque ; `/img` l'exige (pas de repli sur un personnage par défaut) ;\n* toute réponse porte un corps JSON, succès comme échec. Une erreur a la forme\n  `{\"ok\": false, \"erreur\": \"<texte français destiné à l'écran>\"}` ;\n* les uploads voyagent en base64 dans un corps JSON, jamais en\n  `multipart/form-data` — conséquence voulue du garde d'origine.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/img": {
      "get": {
        "summary": "Octets d'une image de tri",
        "description": "Bytes of an image from a sorting bucket. `character=` is MANDATORY here.\n\nThis route serves character data: leaving it a default means serving Léna's\nimages to whoever did not ask for them — which is exactly what happened\nbefore 29/08/2026, where Abyssiaelle's Review screen displayed Léna's\ngallery. A name that is not in the requested character's tree comes out as\n404, never through a fallback onto another tree.\n\n┌── COUPLING TO PRESERVE — migration brief §4.1, AUDIT §5.6.1 ────────────┐\n│ `v` IS DECLARED SO THAT IT IS IGNORED, AND THAT IS THE WHOLE POINT.     │\n│                                                                         │\n│ `v` is the mtime of the bytes, produced by /api/gallery (see            │\n│ routers/review.py, same box) and appended to the URL by the single      │\n│ image-URL builder of the frontend, `imgUrl()` in static/api.js. The     │\n│ server has never read it and must never start: it exists solely to      │\n│ make the URL change when the bytes change.                              │\n│                                                                         │\n│ WHY IT EXISTS. Since the editor learned to overwrite its source         │\n│ (`/api/edit/save?remplacer`, F3.3), one file name can designate two     │\n│ different images. Without `v` in the URL the browser re-serves its      │\n│ cached copy and the screen shows the image from before, on a file that  │\n│ has changed.                                                            │\n│                                                                         │\n│ THREE WAYS TO BREAK IT, all silent — no error, just a stale image:      │\n│   - rejecting the parameter as unknown (a strict query model would);    │\n│   - reading it, and serving from a cache keyed on it;                   │\n│   - changing its format, or emitting it where /api/gallery does not     │\n│     (STATE.recent carries no `v`, and `imgUrl` then omits it — the URL  │\n│     stays character-for-character the one from before).                 │\n│                                                                         │\n│ It is declared here rather than left implicit so /docs states the       │\n│ contract, and so nobody \"cleans up\" an undeclared parameter.            │\n└─────────────────────────────────────────────────────────────────────────┘",
        "operationId": "serve_image_img_get",
        "parameters": [
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "OK, A_REVOIR, REJET, SANS_VISAGE, ARCHIVE",
              "default": "OK",
              "title": "Bucket"
            },
            "description": "OK, A_REVOIR, REJET, SANS_VISAGE, ARCHIVE"
          },
          {
            "name": "space",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "sfw (défaut) ou nsfw",
              "title": "Space"
            },
            "description": "sfw (défaut) ou nsfw"
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Nom de fichier, motif SAFE_NAME",
              "default": "",
              "title": "Name"
            },
            "description": "Nom de fichier, motif SAFE_NAME"
          },
          {
            "name": "thumb",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Non vide : servir la vignette 420×560 au lieu de l'original. N'importe quelle valeur non vide compte, « 0 » y compris.",
              "title": "Thumb"
            },
            "description": "Non vide : servir la vignette 420×560 au lieu de l'original. N'importe quelle valeur non vide compte, « 0 » y compris."
          },
          {
            "name": "v",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Jeton de cache, IGNORÉ par le serveur. Voir la note dans le code : il ne sert qu'à l'URL.",
              "title": "V"
            },
            "description": "Jeton de cache, IGNORÉ par le serveur. Voir la note dans le code : il ne sert qu'à l'URL."
          },
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire.",
              "title": "Character"
            },
            "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire."
          }
        ],
        "responses": {
          "200": {
            "description": "Octets de l'image",
            "content": {
              "image/png": {},
              "image/jpeg": {}
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Absente de l'arbre du personnage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageNotFound"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/img/pose": {
      "get": {
        "summary": "Squelette de pose",
        "description": "Thumbnail of a skeleton from INPUTS/POSE/, for the scene card's pose\npicker. No bucket here — it is a flat folder, not a sorting tree — so not\nthe same path as `serve_image`. Files are small (~50 KB, transparent\nbackground): no resized thumbnail needed.\n\nNo `character=` either, and that is not an oversight: the pose bank is\nshared by every character, like INPUTS/REALISME/ (AUDIT §5.2 lists the two\nglobal remainders).",
        "operationId": "serve_pose_img_pose_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Nom de fichier, motif SAFE_NAME",
              "default": "",
              "title": "Name"
            },
            "description": "Nom de fichier, motif SAFE_NAME"
          }
        ],
        "responses": {
          "200": {
            "description": "Octets de l'image",
            "content": {
              "image/png": {},
              "image/jpeg": {}
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Absente de l'arbre du personnage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageNotFound"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pose/extract": {
      "post": {
        "summary": "Photo -> squelette OpenPose",
        "description": "A photo sent by the user -> an OpenPose skeleton, into INPUTS/POSE/.\n\nThe only web entry point where a real photograph of a third party can\narrive — see AUTOMATION/pose_tools.py, which guarantees it is never\npersisted.\n\nThe extraction runs in an executor: `pose_tools.extraire` talks to ComfyUI\nwith blocking urllib, and a blocking call in an async handler freezes the\nwhole server (the 2005 ms measured on 24/08, see `comfy_alive`).",
        "operationId": "extract_pose_api_pose_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoseExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoseExtractResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "ComfyUI hors ligne"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pose/delete": {
      "post": {
        "summary": "Retirer un squelette",
        "operationId": "delete_pose_api_pose_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoseDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Squelette introuvable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/edit/save": {
      "post": {
        "summary": "Enregistrer une retouche navigateur",
        "description": "Saves a retouch (crop / colour / grain, browser-side).\n\nBy default a NEW file `<name>_edit`, in the same bucket as the original:\nthat is the normal gesture, the original stays comparable and separately\ndeletable through `delete_image`. Neither measurement nor export happens\nautomatically on that path — it is not a generation, and `measure_batch`\nis still there to score the copy.\n\n`remplacer: true` overwrites the source (F3.3, 30/08/2026). The frontend\nonly sends it after an explicit confirmation, and it is never its primary\nbutton. Three consequences are handled HERE, without which the interface\nwould be lying about a file that changed under it:\n\n  - the realism MEASUREMENTS were about the old pixels: erased\n    (`mes.demesurer`), the image becomes « unmeasured » again. The human\n    judgement, however, is kept;\n  - the publishable export (OK/sfw) is redone from the new bytes, otherwise\n    the distributed JPEG stays the image from before;\n  - the thumbnail is forgotten. Its timestamp would be enough to redo it,\n    but `oublier_vignette` makes it certain even if the disk clock is\n    coarser than the gesture.\n\nThe generation's JOURNAL ROW is not touched: it says what the pipeline\nproduced, which stays true — the file simply no longer illustrates it\nexactly. The frontend says so in its confirmation.",
        "operationId": "save_edit_api_edit_save_post",
        "parameters": [
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire.",
              "title": "Character"
            },
            "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditSaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditSaveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Image d'origine introuvable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "erreur": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Erreur"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ActionResponse",
        "description": "The universal answer of an action route, `{ok: bool, erreur?: str}`.\n\nAUDIT §5.4: the convention holds on 4xx and 5xx too — `api.js` reads a JSON\nbody on every response, whatever the status.\n\nALWAYS DECLARE IT WITH `response_model_exclude_unset=True`. The old handlers\nreturned `{\"ok\": true}` and nothing else on success; a plain response_model\nwould add `\"erreur\": null` to every one of them. Harmless to `erreurDe()`,\nwhich only tests `ok === false` — but this migration is meant to keep the\nbodies identical, and a shape that drifts once drifts twice."
      },
...
      "EditSaveRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "bucket": {
            "type": "string",
            "title": "Bucket",
            "default": ""
          },
          "space": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space"
          },
          "remplacer": {
            "type": "boolean",
            "title": "Remplacer",
            "default": false
          },
          "data_base64": {
            "type": "string",
            "title": "Data Base64",
            "default": ""
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "EditSaveRequest",
        "description": "Saves a browser-side retouch (crop / colour / grain).\n\nBy default a NEW file `<name>_edit`, in the same bucket as the original.\n`remplacer: true` overwrites the source (F3.3, 30/08/2026) — the frontend\nonly sends it after an explicit confirmation, and it is never its primary\nbutton."
      },
      "EditSaveResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "remplace": {
            "type": "boolean",
            "title": "Remplace"
          },
          "export": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Export"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "ok",
          "name",
          "remplace"
        ],
        "title": "EditSaveResponse",
        "description": "`remplace` tells the two paths apart: a copy under a new name, or the\nsource overwritten. `export` is only filled on the overwrite path, where\nthe published JPEG had to be redone from the new bytes."
      },
...
      "ErrorResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": false
          },
          "erreur": {
            "type": "string",
            "title": "Erreur"
          }
        },
        "type": "object",
        "required": [
          "erreur"
        ],
        "title": "ErrorResponse",
        "description": "What every rejection looks like. `erreur` is French: it is displayed to\nthe user verbatim (AUDIT §5.5)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageNotFound": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": false
          },
          "erreur": {
            "type": "string",
            "title": "Erreur"
          }
        },
        "type": "object",
        "required": [
          "erreur"
        ],
        "title": "ImageNotFound",
        "description": "404 of a byte-serving route.\n\nA blunt 404 on purpose: the file is not in THIS character's tree. Never a\nlookup in another tree, even if the name exists there — that fallback was\nthe isolation bug of 29/08/2026.\n\nUnder aiohttp this came out as bare text (`raise web.HTTPNotFound()`); it is\nJSON now, like every other response (AUDIT §5.4). Nothing reads it: these\nURLs sit in `<img src>`."
      },
...
      "PoseDeleteRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "PoseDeleteRequest",
        "description": "Removes a skeleton from the bank. Touches nothing else: a scene still\nreferencing it loses it silently at file level — `validate_scene_bank` will\nreport it on the next save of scenes.json (« squelette introuvable »)."
      },
      "PoseExtractRequest": {
        "properties": {
          "data_base64": {
            "type": "string",
            "title": "Data Base64",
            "default": ""
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "default": "photo.png"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "PoseExtractRequest",
        "description": "A photo sent by the user, to be turned into an OpenPose skeleton.\n\nTHE ONLY WEB ENTRY POINT WHERE A REAL PHOTOGRAPH OF A THIRD PARTY CAN\nARRIVE. `AUTOMATION/pose_tools.py` guarantees it is never persisted: only\nthe skeleton reaches the disk.\n\nbase64 inside a JSON body, never `multipart/form-data` — see api/security.py\nfor why that is a consequence of the origin guard and not an oversight. The\nceiling is `ss.TAILLE_MAX_PHOTO` (20 MB before encoding)."
      },
      "PoseExtractResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "name"
        ],
        "title": "PoseExtractResponse",
        "description": "`name` is the skeleton written to INPUTS/POSE/, which a scene may then\nreference by that file name."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}